requirements.txt
opencv-python
numpy
requests
pyserial
flask
python-dotenv
onnxruntime
insightface
sqlalchemy
streamlit
streamlit-autorefresh
discord.py
pandas
colorama
